Updated Generator
Added support for Emoticon file to be able to specify all the emoticons
on a single file rather than inside the code. There is no need to change
the code now if there is a need to add more emoticons. This also makes
it easier to reorganize the emoticons for creating a nicer emoji art.